zookeeper and curator added and migrated to java 11

+++ backend-service/src/main/java/org/gbif/occurrence/annotation/controller/ProjectController.java
@@ -13,8 +13,6 @@
  */
 package org.gbif.occurrence.annotation.controller;
 
-import io.swagger.v3.oas.annotations.tags.Tag;
-
 import org.gbif.occurrence.annotation.mapper.ProjectMapper;
 import org.gbif.occurrence.annotation.mapper.RuleMapper;
 import org.gbif.occurrence.annotation.mapper.RulesetMapper;
@@ -31,6 +29,7 @@ import org.springframework.web.bind.annotation.*;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.tags.Tag;
 
 import static org.gbif.occurrence.annotation.controller.AuthAdvice.assertCreatorOrAdmin;
 
@@ -50,8 +49,8 @@ public class ProjectController implements Controller<Project> {
   public List<Project> list(
       @RequestParam(required = false) Integer limit,
       @RequestParam(required = false) Integer offset) {
-    int limitInt = limit == null ? 100 : limit.intValue();
-    int offsetInt = offset == null ? 0 : offset.intValue();
+    int limitInt = limit == null ? 100 : limit;
+    int offsetInt = offset == null ? 0 : offset;
     return projectMapper.list(limitInt, offsetInt);
   }
