+++ backend-service/src/main/java/org/gbif/occurrence/annotation/controller/RulesetController.java
@@ -13,8 +13,6 @@
  */
 package org.gbif.occurrence.annotation.controller;
 
-import io.swagger.v3.oas.annotations.tags.Tag;
-
 import org.gbif.occurrence.annotation.mapper.RuleMapper;
 import org.gbif.occurrence.annotation.mapper.RulesetMapper;
 import org.gbif.occurrence.annotation.model.Ruleset;
@@ -30,6 +28,7 @@ import org.springframework.web.bind.annotation.*;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.tags.Tag;
 
 import static org.gbif.occurrence.annotation.controller.AuthAdvice.assertCreatorOrAdmin;
 
@@ -50,8 +49,8 @@ public class RulesetController implements Controller<Ruleset> {
       @RequestParam(required = false) Integer projectId,
       @RequestParam(required = false) Integer limit,
       @RequestParam(required = false) Integer offset) {
-    int limitInt = limit == null ? 100 : limit.intValue();
-    int offsetInt = offset == null ? 0 : offset.intValue();
+    int limitInt = limit == null ? 100 : limit;
+    int offsetInt = offset == null ? 0 : offset;
     return rulesetMapper.list(projectId, limitInt, offsetInt);
   }
